manager: DRING_ID_MAX_VAL is uint64_t

+++ src/manager.h
@@ -61,7 +61,7 @@ class IceTransportFactory;
 class DataTransferFacade;
 class JamiAccount;
 
-static constexpr size_t DRING_ID_MAX_VAL = 9007199254740992;
+static constexpr uint64_t DRING_ID_MAX_VAL = 9007199254740992;
 
 /** Manager (controller) of daemon */
 class Manager {
